requirements.txt
requests
python-dotenv
openai
flask
gunicorn